poster-scheduler/src/requirements.txt
google-api-core
google-auth
google-cloud-core
google-cloud-secret-manager
google-cloud-storage
google-cloud-vision
instagrapi
Pillow
pytest
geopy
moviepy
pilmoji
tenacity
PyYaml
